pyproject.toml
[project]
name = "django-starter"
version = "0.1.0"
description = "Start Project for Django"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.9",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "ipython>=9.7.0",
    "whitenoise>=6.11.0",
]
